Moved hideCity and RentPlot to their own files. weather renamed as it was not a plot
