+++ src/components/pages/CompareCities/PlotlyHelpers/weather.js
